GCD is added (needs to be rebuilt)

+++ src/OpDiv.java
@@ -1,6 +1,8 @@
 public class OpDiv implements Operation{
     @Override
-    public double exec(double a, double b) {
+    public int exec(int a, int b) {
         return a / b;
     }
 }
+
+
