+++ src/OpMinus.java
@@ -1,9 +1,6 @@
-/**
- * Created by AnastasiyaPC on 7/31/2017.
- */
 public class OpMinus implements Operation{
     @Override
-    public double exec(double a, double b) {
+    public int exec(int a, int b) {
         return a - b;
     }
 }
