+++ src/OpMul.java
@@ -1,6 +1,6 @@
 public class OpMul implements Operation{
     @Override
-    public double exec(double a, double b) {
+    public int exec(int a, int b) {
         return a * b;
     }
 }
